refactor account information header as separated component
