+++ apps/web/src/components/AccountHeader.styles.tsx
@@ -1,0 +1,8 @@
+import styled from "@emotion/styled";
+
+export const Root = styled.div`
+    width: 100%;
+
+    display: flex;
+    align-items: center;
+`;
