+++ apps/web/src/components/Column/Base.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { mergeRefs } from "react-merge-refs";
 import Scrollbars from "rc-scrollbars";
 
-import { Avatar, IconButton } from "ui";
+import { IconButton } from "ui";
 
-import { Box, LinearProgress, Stack, SvgIconProps, Typography } from "@mui/material";
+import { LinearProgress, Stack } from "@mui/material";
 import SettingsRoundedIcon from "@mui/icons-material/SettingsRounded";
 import DeleteRoundedIcon from "@mui/icons-material/DeleteRounded";
 
@@ -14,13 +14,13 @@ import { ColumnSidebarProps } from "@components/Column/Sidebar/Base";
 import { ColumnSettingsSidebar } from "@components/Column/Sidebar/Settings/Column";
 import { ColumnContext } from "@components/Column/context";
 import { ColumnInstance } from "@components/Column/types";
-import { MastodonLogo } from "@components/Svg/Mastodon";
 import { Content, Handle, Header, ProgressWrapper, Root, Title, Wrapper } from "@components/Column/Base.styles";
 
 import { BaseAccount } from "@services/base/account";
 
 import { useColumnNodeSetter, useColumns } from "@states/columns";
 import { Fn } from "@utils/types";
+import { AccountHeader } from "@components/AccountHeader";
 
 export interface ColumnProps {
     instance: ColumnInstance;
@@ -33,10 +33,6 @@ export interface ColumnProps {
 interface SidebarHolder {
     component: React.ComponentType<ColumnSidebarProps> | null;
 }
-
-const SERVICE_TYPE_TO_ICON: Record<string, React.ComponentType<SvgIconProps>> = {
-    mastodon: MastodonLogo,
-};
 
 export const BaseColumn = ({ instance, children, loading, onScroll, account }: ColumnProps) => {
     const { id, title } = instance;
@@ -75,15 +71,6 @@ export const BaseColumn = ({ instance, children, loading, onScroll, account }: C
         sidebarContent = <Component instance={instance} />;
     }
 
-    let headerLogo: React.ReactNode | null = null;
-    if (account) {
-        const serviceType = account.getServiceType();
-        const ServiceIcon = SERVICE_TYPE_TO_ICON[serviceType];
-        if (ServiceIcon) {
-            headerLogo = <ServiceIcon fontSize="inherit" />;
-        }
-    }
-
     return (
         <ColumnContext.Provider value={{ column: instance }}>
             <Wrapper style={{ ...style, zIndex: isDragging ? 1000 : 0 }}>
@@ -98,37 +85,7 @@ export const BaseColumn = ({ instance, children, loading, onScroll, account }: C
                     <Header>
                         <Handle {...attributes} {...listeners} />
                         <Title role="button" onClick={handleTitleClick}>
-                            {account && (
-                                <Box mr={1}>
-                                    <Avatar size="small" src={account.getAvatarUrl()} />
-                                </Box>
-                            )}
-                            <Box minWidth={0} display="flex" flexDirection="column" justifyContent="center">
-                                <Typography
-                                    variant="h6"
-                                    fontSize="0.95rem"
-                                    fontWeight={600}
-                                    lineHeight={1}
-                                    whiteSpace="nowrap"
-                                    overflow="hidden"
-                                    textOverflow="ellipsis"
-                                    sx={{ mb: account ? 0.5 : 0 }}
-                                >
-                                    {title}
-                                </Typography>
-                                {account && (
-                                    <Typography
-                                        variant="body2"
-                                        fontSize="0.8rem"
-                                        color="text.secondary"
-                                        whiteSpace="nowrap"
-                                        overflow="hidden"
-                                        textOverflow="ellipsis"
-                                    >
-                                        {headerLogo} {account.getDisplayName()}
-                                    </Typography>
-                                )}
-                            </Box>
+                            <AccountHeader avatarSize="small" account={account} titleText={title} />
                         </Title>
                         <Stack direction="row" spacing={1} className="controls" flex="0 0 auto" ml={1.5}>
                             <IconButton
